Start gathering execution statistics
These are mostly about the major steps taken during a build. More to come.
Also, the rendering thus far is pathetic — I'll fix that up soon. Also, now
that we have the information at program termination, it's easy to signal
successful/unsuccessful termination to the caller. This is an important feature
to make this tool useful in the realm of continuous integration.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>