Use tools from configuration unless a module overrides
It's not useful to list toolchains over and over again. It's better to list
requirements for a toolchain and then use all toolchains from a pool in con-
figuration that can satisfy them.

Modules can still override this broad list. Which is useful with modules that
are actually firmwares that only make sense with a fixed set of build tools.

Same goes for buildconfigs and buildtools — though at this point they don't
check for any dependencies.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>