Add output policy options
These are not yet used. The logic is that the two booleans invert the meaning
of log-all and log-to-file after loading configuration files. That means that
users can change the default for themselves and still get the opposite opera-
tion via command-line options. The defaults in config.yaml are ‘false’ for
both.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>