Implement to-file-logging
Since we're deciding about destination file names and so on, during the
program's startup, the default logbook handlers don't quite cut it.

Therefore, this adds a custom handler, that buffers incoming records until it's
told a destination. It then drains the backlog into that destination and then
resumes the usual job of a logbook handler.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>