Implement the missing top-level commands
- build-tree-init
- fetch-dependencies
- generate-toplevel
- run-instance

These may not be perfect yet. It's certainly missing the handling of
--log-to-file. Also, this is quite a bit of code-duplication, which
should be refactored for obvious reasons.

Now, however, you can do this:

  mmh -d ci build
  mmh -d ci run

And use "run" all over again without having to remove "ci". This allows expe-
rimenting with changes to the generated toplevel cmake lists file as well as
the contents of "deps".

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>