Experimental: Allow ‘run’ to re-run selected instances
This allows:

  mmh get ufw
  cd ufw
  mmh -d ci -T gnu-arm-none-eabi -A cortex-m3 -B ninja -C debug,release
  mmh -d ci run ./ci/build/gnu-arm-none-eabi_cortex-m3_none_debug_ninja

This *only* re-runs the selected build instances in a pristine build directory.
While this is a useful way of doing things, it has a few draw-backs:

- Does not re-use statistics-gathering sub-system.
- Could be better abstracted and share more code with the existing building
  code.
- Requires specifying ‘-d’.

The first two points would yield a cleaner code-base. The third could be impro-
ved upon by adding the build-root directory's name in MakeMeHappy.yaml and then
walk up the directory-tree to find that file and re-run accordingly.

That way you could enter a build directory and issue ‘mmh run’ to re-run. No
options required. It would also lift the requirement to define ‘-d’ when speci-
fying instance directories to rebuild on the command line.

Therefore, this is marked experimental for now. It should be possible to
support the current call structure in the improved system, too. But no pro-
mises.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>