Preliminary support for building zephyr applications
This can build zephyr applications such as blinky, when provided with a
module.yaml file such as this:

    name: zephyr-sample-blinky
    type: zephyr

    dependencies:
    - name: zephyr-kernel
        revision: v2.6.0
    - name: ufw
        revision: master
    - name: zephyr-arm-cmsis
        revision: c3bd2094f92d574377f7af2aec147ae181aa5f8e
    - name: zephyr-hal-stm32
        revision: d49fb6dbe780b931eee33cce8656e87c46070fa8

    targets:
    - board: stm32h747i_disco_m7
        toolchain: gnuarmemb
        modules: [ zephyr-arm-cmsis, zephyr-hal-stm32 ]
    - board: nucleo_f767zi
        toolchain: gnuarmemb
        modules: [ zephyr-arm-cmsis, zephyr-hal-stm32 ]

    buildconfigs:
    - release
    - debug

    buildtools:
    - ninja

    defaults:
    GNUARMEMB_TOOLCHAIN_PATH: /usr

The ufw dependency is mandatory, even if the application does not use ufw
itself. We're using lots of ufw's zephyr integration to smooth over some of
zephyr's peculiarities when it comes to CMake usage.

Not sure if this is the last word with respect to zephyr support, but it is a
decent start.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>