build: Allow zephyr instances to define application key
If your zephyr application does not live in the root directory of a repository,
zephyr's kconfig system will miss a couple of possible configuration files,
since the APPLICATION_SOURCE_DIR parameter points to the wrong directory. This
allows users to set this to a sub-directory within the code-under-test link.
This also allows a repository to contain multiple zephyr applications, and all
this can be handled by multiple build instances that configure this parameter
to different values.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>