Add --show-phases option
When using --log-to-file, mmh does not provide a lot of feedback to the user.
That's sort of the point of course. But since mmh can be very useful in
continuous integration environments, where --log-to-file is useful not to
overwhelm the webinterface or to allow users to download the full build log, it
might still be nice to get a little progress feedback in console output.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>