Add top-level --succeed option
The idea here is to be able to force the return value of the command to suc-
cess, if possible. Broken option specs will still fail. The purpose is to be
able to do things like this in CI:

    setup-ci-sequence
    mmh --succeed --log-to-file --show-phases --log-file build.log
    teardown-ci-sequence
    mmh --quiet result build.log

That way you can inspect the build result at the very end of CI output. Any
output from ‘teardown-ci-sequence’ will not obscure output.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>