zephyr: Inherit west.yml revisions for revision-less dependencies
This is a good step in order to make switching to different zephyr kernels
easy.

Source stack lookup leaves something to be desired: When a higher priority file
changes, say the repository-url of a module, that causes lookup() to forget
useful things like the 'west' mapping. This information should be inherited,
from lower priority files if they have it.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>