source stack: Add default 'main' parameter
This refers to the main development branch of a project. If it is a list of
strings, the first entry that exists within the repository wins. If it is a
string, only that string is considered. The default is main and master, which
should be good enough for most projects out there. It is still configurable for
outliers, that I am sure exist.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>