Implement use-latest-revision override spec
The way this works, is that mmh checks out the specified main branch of the
modules that match the override pattern. After that if finds the closest tag
visible from that branch and uses it as the revision of the module.

It is possible to constrict the set of possible tags by specifying a file
globbing like pattern in use-latest-revision-pattern. This can also be achieved
from the command line by doing:

    mmh --revision 'foobar=!latest:v*'

This only allows tags that match "v*" to be used. The ":PATTERN" part is
optional and the default pattern is "*".

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>