Add a hook sub system
This uses the "evaluate" key to inject python code into the operation of mmh.

To try this out, you can use this in module.yaml and system.yaml:

    evaluate: foobar.py

With foobar.py containing this:

    import makemehappy.hooks as h

    def myhook(kind, **kwargs):
        print('myhook:', f'kind:{kind}', kwargs)

    h.hooks.register('.', myhook)

And that'll print all hooks being run with the available arguments presented to
them. The first argument to "register()" is a pattern, that is matched against
the "kind" property of the specific hook. The "." regex matches every hook
kind, thus running myhook() with each and every hook.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>