Allow system to specify build-combinations
Sometimes when building complete systems, to achieve finally useable results
require to combine the results of other build-steps. For instance, you may want
to create flashable images that contain a bootloader and an application, both
of which were build in basic build steps.

To do this, combinations allow specifying a list of parents (both if which must
be built), after which a combination runner callback is invoked so the user can
inject arbitrary code to perform the combination.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>