deploy: Make path-like classes portable across python versions
Trying to run "mmh system deploy" at work flamed our gloriously, since the de-
bian stable (bookworm) systems ship python 3.11.x, in which the classes from
pathlib are not subclassable. Many people patched some of the internals when
trying to derive from things like Path, to make this sort of work.

Some details about this can be found here:

  https://discuss.python.org/t/make-pathlib-extensible/3428

This effort ended in change in python 3.12.x's pathlib, where deriving from
these classes is now supported like it is with all usual classes. But the
downside is that the workarounds for deriving in older versions now stopped
working. Great… not that the change is unwelcomed. It's actually awesome and
makes using these classes so much more natural. But doing things portably is
now a bit of a nightmare.

So for the time being, this feature now does not derive its path-like classes
derive from pathlib.Path. Instead this adds a Path member to all these classes
and implements enough of the Path API so almost no changes to the rest of the
code are needed.

I hope that in a year or two python 3.11.x and older support can be dropped,
when debian bookworm gets phased out. Then this commit should be reverted.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>