Implement checksum calculation on top of hashlib
This is at three levels: A file→hexsum function; on top of that a function that
does file→sum-like record; and on top of that a function that basically imple-
ments what programs like md5sum, sha256sum, and sha512sum do. For these three I
checked, and the output is compatible. So we can use this during deployment to
add artefact consistency.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>