Add a simple verify command
This allows checking the checksum files generated by system deploy. Well,
really it allows any of md5, sha256, and sha512 checksum files. But it is
geared toward verifying what system deploy built. In particular, it allows
being called with a list of directories and it will look for those files in
these directories and attempt to verify their contents with file root pre-
fixes based on the name of the directory handed in. Example:

  % mmh verify ~/web/releases/foobar-v0.1.23

…instead of:

  % cd ~/web/releases/foobar-v0.1.23
  % md5sum contents.md5
  % sha256sum contents.sha256
  % sha512sum contents.sha512

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>