Make system build run rebuild if it can
Incremental builds are a good thing. This will create a stamp file whenever a
configure process succeeds. Then, if a build runs and this file exists, the
build is not rerun completely, but it defers to rebuild so the configure phase
gets skipped. Whenever a full build is triggered, existing stamp files are
removed before running the configure phase so the configure phase must succeed
whenever it runs.

The --force-build (-f) option to "system build" forces the configure phase
irregardless of the stamp file's existence.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>