sys: Add combination build support
When mmh is used to build more complex systems, for instance some that build
both bootloaders and firmware for certain applications, it is desirable to
generate additional artefacts by combining the artefacts from previous build
results. This can generate images with added meta-data, signed images, or
bootstrap-images that are the combination of a bootloader and a firmware for
initial provisioning of a system.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>