Add a way to track the state of mmh the program
When combination builds will come around, it is useful if mmh the program
changed from one artefact generation to the next, because the changed code
might change the artefact generation.

This works by checksumming code of loaded modules and code of the mmh program
and the code of any python snippets loaded at runtime. This tries to cache
these results a well as possible.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>