Remove build/MakeMeHappy.yaml and the reset-setup sub-command
On a whole, the use of MakeMeHappy.yaml was determental to mmh's use. The plan
was to allow users to have to do less redundant specifications on the command
line. But in reality, it is confusing to most. It also made the initialition
code for build directories much more complex than it should be. And finally, it
was the source of subtle bugs like:

 % mmh system build zephyr/*/chip-remote/gnuarmemb/debug
 % mmh system db zephyr/nucleo_f302r8/chip-remote/gnuarmemb/debug
 % mmh system build

With the latter command only building the instance named in the "db" call.

This removes these issues by removing the file completely. It also removes the
reset-setup sub-command, since its purpose was to reset MakeMeHappy.yaml con-
tents to its defaults.

Now, it would still be useful to be able to tell if a directory is used as a
build directory by mmh, and to distinguish the mode used by mmh (module build,
system build, single-instance system mode). This should be done with minimal
structure, and once established, this state should be read-only. It will hap-
pen again at some point.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>