comb: Make CombinationOutput names more safe
This disallows the creation of dot-files with this system, since we're using
dot-files for tracking output state, and this will avoid possible clashes. This
also disallows "mmh-state" as an output name, since that would lead to a state
file named ".mmh-state.yaml" — but that file is already used to track the state
of the combination the output file is a part of.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>