Refactor main program
The main program was always a little ugly with a lot of stuff being copy and
pasted just enough to work kind of right, even though the toplevel command set
was always made to complement each other.

This puts the CLI parser into makemehappy.cli and the bulk of the main program
into makemehappy.program, while removing a lot of the redundant code. Not all
possible calling structures were tested yet, but the most common things work.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>