comb: Improve handling of dependency-less combinations
When solely specified on the command line, these would cause the whole system
to be built, because the list of activated instances would be zero, and so the
system would default to building everything.

That was also a deficiency: When a combination is specifed, all its dependen-
cies are built and during that, all possible combinations are being executed,
rather than the ones actually specified by the user. The rectified this as
well.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>