zephyr: Define ZEPHYR_BASE in execution environment
There is some code in Zephyr (in samples, for instance), that uses ZEPHYR_BASE
from the execution environment to address some source files, and if it is not
set, that'll break, obviously.

Personally, I think they should be using ${ZEPHYR_BASE} everywhere to figure
out file names. But there is zephyr-env.sh in the repository root that would
set this environment variable, so maybe that's intentional.

At any rate, it is fairly straight forward to make mmh set this parameter as
well when it invokes cmake for Zephyr purposes. That is what this patch does.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>