Fix number-of-expected-instances with combinations without dependancies
The countPossible() method returned the number of combinations that could be
build given the active set of builds. When a combination without any depen-
dencies was registered, that always counted as a possible build.

When the whole system is built, that is correct. But if builds are explicitly
specified on the command line, those should be disregarded, unless they them-
selves are among those explicitly specified builds.

This accomplishes that by not considering builds that are already built as
possible. Then with only() marking everything that is not explicitly specified
as done, the goal is met.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>