Fix module count in download-sources
The code was resetting the expected number of modules for each file in the
sources stack. Note on the implementation: We're not using .merged here, so we
can provide more detailed information about where a definition originated, in
case an error occurs. This could be achieved differently also, but this works.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>