download-sources: Allow filtering sources with regular expressions
There are use cases, for instance in CI, where you want to be able to limit the
sites you want your package system to allow to fetch from. For example, it
might be a policy only to fetch from mirrors on the premises of a company
instead of the whole of the internet. With these patterns, that is possible to
limit.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>