combinations: Fix output if part of combination build fails
If a part of a combination build fails mmh tries to log the
failed ParentInstance's description field which doesn't exist.
As the part of the concatonated string which tries to access the
description member was not an f-string python didn't evaluate it so
it didn't throw an error.

Make the string an f-string and use the name field instead.